UI Improvements
-Added header in Message panel.
-Moved gems text and icon to message panel.

+++ Assets/Scripts/MVC/ChestModel.cs
@@ -10,7 +10,7 @@ public class ChestModel
     public int gems { get;   }
     public string gemsRange { get; }
     
-    public int unlockTime { get; set; }
+    public float unlockTime { get; set; }
     public int GemsRequiredToUnlock { get; set; }
     
     public Sprite lockedChestSprite { get; }
